% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.pack-package.R
\docType{package}
\name{f.pack}
\alias{f.pack}
\title{f.pack: plots para testes de hipóteses.}
\description{
O pacote f.pack gera a representação gráfica para interpretação de um
teste de hipóteses bilateral para a média de uma população
}
\section{Funções do pacote}{

\code{\link[f.pack]{t_calc}}
\code{\link[f.pack]{z.calc}}
\code{\link[f.pack]{t_crit}}
\code{\link[f.pack]{z.crit}}
\code{\link[f.pack]{t_hip}}
\code{\link[f.pack]{z.hip}}
}

