% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc1.R
\name{t_calc}
\alias{t_calc}
\title{Calcula o valor da estatística para um teste de hipóteses bilateral sob a distribuição t de Sudent
amostras de reduzido tamanho extraídas de uma população reconhecidamente Normal}
\usage{
t_calc(dados, H0)
}
\arguments{
\item{dados}{O vetor com os elementos amostrais extraídos da população sob teste}

\item{H0}{O valor assumido sob a hipótese nula}
}
\description{
Calcula o valor da estatística para um teste de hipóteses bilateral sob a distribuição t de Sudent
amostras de reduzido tamanho extraídas de uma população reconhecidamente Normal
}
\examples{
\dontrun{
dados=c(10,11,12,13)
H0=11
t_calc(dados,H0)
}
}
