% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc1.R
\name{t_crit}
\alias{t_crit}
\title{Calcula o valor crítico para a estatística para um teste de hipóteses bilateral sob a distribuição t de Sudent para
amostras de reduzido tamanho extraídas de uma população reconhecidamente Normal}
\usage{
t_crit(alfa, dados)
}
\arguments{
\item{alfa}{O valor do nível de significância arbitrado}

\item{dados}{O vetor com os elementos amostrais extraídos da população sob teste}
}
\value{
O quantil crítico da distribuição t de Student para o nível de significância arbitrado
}
\description{
Calcula o valor crítico para a estatística para um teste de hipóteses bilateral sob a distribuição t de Sudent para
amostras de reduzido tamanho extraídas de uma população reconhecidamente Normal
}
\examples{
\dontrun{
dados=c(10,11,12,13)
alfa=0.05
t_crit(alfa, dados)
}
}
\seealso{
\url{http://www.r-pŕoject.org}
para mais informações sobre a função \code{\link[stats]{qt}}
}
