% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc1.R
\name{t_hip}
\alias{t_hip}
\title{Plota o distribuição t de Student com todos os elementos para o teste de hipóteses bilateral para
amostras de reduzido tamanho extraídas de uma população reconhecidamente Normal}
\usage{
t_hip(alfa, dados, H0)
}
\arguments{
\item{alfa}{O nível de significância arbitrado}

\item{dados}{O vetor com os dados da amostra extraida da população sob teste}

\item{H0}{O valor assumido sob a hipótese nula}
}
\description{
Plota o distribuição t de Student com todos os elementos para o teste de hipóteses bilateral para
amostras de reduzido tamanho extraídas de uma população reconhecidamente Normal
}
\examples{
\dontrun{
dados=c(10,11,12,13)
alfa=0.05
H0=11
t_hip(alfa, dados, H0)
}
}
