% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc1.R
\name{z.calc}
\alias{z.calc}
\title{Calcula o valor da estatística para um teste de hipóteses bilateral sob a distribuição Normal padronizada}
\usage{
z.calc(dados, H0, sigma)
}
\arguments{
\item{dados}{O vetor com os elementos amostrais extraídos da população sob teste}

\item{H0}{O valor assumido sob a hipótese nula}

\item{sigma}{Desvio padrao populacional conhecido ou arbitrado}
}
\description{
Calcula o valor da estatística para um teste de hipóteses bilateral sob a distribuição Normal padronizada
}
\examples{
dados=c(10,11,12,13)
H0=11
sigma=2
z.calc(dados, H0, sigma)
}
