% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc1.R
\name{z.crit}
\alias{z.crit}
\title{Calcula o valor crítico para a estatística para um teste de hipóteses bilateral sob a distribuição Normal padronizada}
\usage{
z.crit(alfa)
}
\arguments{
\item{alfa}{O valor do nível de significância arbitrado}
}
\value{
O quantil crítico da distribuição Normal padronizada para o nível de significância arbitrado
}
\description{
Calcula o valor crítico para a estatística para um teste de hipóteses bilateral sob a distribuição Normal padronizada
}
\examples{
alfa=0.05
z.crit(alfa)
}
\seealso{
\url{http://www.r-pŕoject.org}
para mais informações sobre a função \code{\link[stats]{dnorm}}
}
