% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc1.R
\name{z.hip}
\alias{z.hip}
\title{Recebe o vetor com os dados amostrais, o nível de significância e o valor assumido pela hipótese nula}
\usage{
z.hip(alfa, dados, H0, sigma)
}
\arguments{
\item{alfa}{O nível de significância arbitrado}

\item{dados}{O vetor com os dados da amostra extraida da população sob teste}

\item{H0}{o valor assumido sob a hipótese nula}

\item{sigma}{Desvio padrao populacional conhecido ou arbitrado}
}
\description{
Recebe o vetor com os dados amostrais, o nível de significância e o valor assumido pela hipótese nula
}
\examples{
dados=c(10,11,12,13)
alfa=0.05
H0=11
sigma=2
z.hip(alfa, dados, H0, sigma)
}
